# nocturnal-archive-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nocturnal Archive API
  description: |
    Enterprise-grade academic research API with numeric grounding and finance analysis.
    
    ## Papers API (Demo)
    Public demo version with limited functionality. For commercial use, see FinSight vertical.
    
    ## FinSight API (Commercial)
    Advanced finance analysis with SEC filing retrieval, numeric grounding, and claim verification.
    
    ## Authentication
    All endpoints require API key authentication via `X-API-Key` header or `Authorization: Bearer <key>`.
    
    ## Error Handling
    All errors follow RFC 7807 Problem+JSON format with machine-readable error types.
    See [Error Catalog](problem_catalog.md) for complete error reference.
  version: 1.0.0
  contact:
    name: Nocturnal Archive Support
    email: contact@nocturnal.dev
    url: https://nocturnal.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://nocturnal.dev/legal
  x-legal-disclaimer: |
    This service is not investment advice. See [Legal Disclaimers](legal_disclaimers.md) for complete terms.
servers:
  - url: https://api.nocturnal.dev
    description: Production server
  - url: http://localhost:8000
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/api/papers/search:
    post:
      summary: Search academic papers (Demo)
      description: |
        Search academic papers with basic functionality.
        
        **Demo Limitations:**
        - Maximum 50 results per request
        - Basic search only
        - Rate limited to 100 requests per month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PapersSearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limit exceeded
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/api/papers/synthesize:
    post:
      summary: Synthesize academic papers (Demo)
      description: |
        Synthesize academic papers with basic functionality.
        
        **Demo Limitations:**
        - Maximum 500 words per summary
        - Maximum 10 papers per synthesis
        - Rate limited to 100 requests per month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PapersSynthesizeRequest'
      responses:
        '200':
          description: Synthesis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SynthesisResult'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          description: Rate limit exceeded
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  # Finance API (Commercial)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for authentication
  schemas:
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: Error type URI
          example: "https://nocturnal.dev/errors/validation-error"
        title:
          type: string
          description: Error title
          example: "Validation error"
        status:
          type: integer
          description: HTTP status code
          example: 422
        detail:
          type: string
          description: Error detail
          example: "The request data is invalid"
        instance:
          type: string
          description: Request ID for correlation
          example: "req_12345"
      required: [type, title, status]

    # Papers API schemas
    PapersSearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Search query
          example: "machine learning"
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
          description: Number of results (max 50 for demo)
        sources:
          type: array
          items:
            type: string
          default: ["openalex"]
          description: Data sources
        year_from:
          type: integer
          description: Start year filter
        year_to:
          type: integer
          description: End year filter
      required: [query]
    PapersSynthesizeRequest:
      type: object
      properties:
        paper_ids:
          type: array
          items:
            type: string
          description: Paper IDs to synthesize
        max_words:
          type: integer
          minimum: 50
          maximum: 500
          default: 200
          description: Maximum words (max 500 for demo)
        style:
          type: string
          default: "bullet_points"
          description: Summary style
      required: [paper_ids]
    SearchResult:
      type: object
      properties:
        papers:
          type: array
          items:
            $ref: '#/components/schemas/Paper'
        count:
          type: integer
          description: Number of results
        query_id:
          type: string
          description: Query identifier
        trace_id:
          type: string
          description: Request trace ID
    SynthesisResult:
      type: object
      properties:
        summary:
          type: string
          description: Generated summary
        paper_ids:
          type: array
          items:
            type: string
          description: Synthesized paper IDs
        word_count:
          type: integer
          description: Summary word count
        trace_id:
          type: string
          description: Request trace ID
    Paper:
      type: object
      properties:
        id:
          type: string
          description: Paper ID
        title:
          type: string
          description: Paper title
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        year:
          type: integer
          description: Publication year
        doi:
          type: string
          description: DOI URL
        abstract:
          type: string
          description: Paper abstract
        citations_count:
          type: integer
          description: Number of citations
        venue:
          type: string
          description: Publication venue
        source:
          type: string
          description: Data source
    Author:
      type: object
      properties:
        name:
          type: string
          description: Author name
        orcid:
          type: string
          description: ORCID ID
        affiliation:
          type: string
          description: Author affiliation

    # Finance API schemas
...